requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.0
django-filter==23.3
django-allauth==0.57.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
Pillow==10.1.0
channels==4.0.0
channels-redis==4.1.0
django-notifications-hq==1.8.3
django-ratings==0.1.0
django-location-field==2.7.3
django-celery==5.3.1
celery==5.3.4
redis==5.0.1
geopy==2.4.0
drf-yasg==1.21.7 